Base Functions

CoTETE.calculate_TE_from_event_times — Function

function calculate_TE_from_event_times(
    target_events::Array{<:AbstractFloat},
    source_events::Array{<:AbstractFloat},
    l_x::Integer,
    l_y::Integer;
    auto_find_start_and_num_events::Bool = true,
    start_event::Integer = 1,
    num_target_events::Integer = length(target_events) - start_event,
    num_samples_ratio::AbstractFloat = 1.0,
    k_global::Integer = 5,
    conditioning_events::Array{<:AbstractFloat} = [0.0],
    l_z::Integer = 0,
    metric::Metric = Euclidean(),
    is_surrogate::Bool = false,
    surrogate_num_samples_ratio::AbstractFloat = 1.0,
    k_perm::Integer = 5,
    )

Estimates the TE from lists of raw event times.

Examples

julia> source = sort(1e4*rand(Int(1e4)));

julia> target = sort(1e4*rand(Int(1e4)));

julia> using CoTETE

julia> TE = CoTETE.calculate_TE_from_event_times(target, source, 1, 1)
0.0

julia> abs(TE - 0) < 0.02 # For Doctesting purposes
true

Arguments

target_events::Array{<:AbstractFloat}: A list of the raw event times in the target process. Corresponds to $X$ in [1].

source_events::Array{<:AbstractFloat}: A list of the raw event times in the source. Corresponds to $Y$ in [1].

l_x::Integer: The number of intervals in the target process to use in the history embeddings. Corresponds to $l_X$ in [1].

l_y::Integer: The number of intervals in the source process to use in the history embeddings. Corresponds to $l_Y$ in [1].

auto_find_start_and_num_events::Bool = true: When set to true, the start event will be set to the first event for which there are sufficient preceding events in all processes such that the embeddings can be constructed. The number of target events will be set such that all time between this first event and the last target event is included.

start_event::Integer = 1: only used when auto_find_start_and_num_events = false. The index of the event in the target process from which to start the analysis.

num_target_events::Integer = length(target_events) - start_event: only used when auto_find_start_and_num_events = false. The TE will be calculated on the time series from the timestamp of the start_event-th event of the target process to the timestamp of the start_event + num_target_events-th event of the target process.

num_samples_ratio::AbstractFloat = 1.0: Controls the number of samples used to estimate the probability density of histories unconditional of the occurrence of events in the target process. This number of samples will be num_samples_ratio * num_target_events. Corresponds to $N_U/N_X$ in [1].

k_global::Integer = 5: The number of nearest neighbours to consider in initial searches.

conditioning_events::Array{<:AbstractFloat} = [0.0]: A list of the raw event times in the target process. Corresponds to $Z_1$ in [1].

Single conditioning process

Note that the framework developed in out paper [1] considers an arbitrary number of extra conditioning processes, at present the framework can only handle a single such process. This will change in future releases.

l_z::Integer = 0: The number of intervals in the single conditioning process to use in the history embeddings. Corresponds to $l_{Z_1}$ in [1].

metric::Metric = Euclidean(): The metric to use for nearest neighbour and radius searches.

is_surrogate::Bool = false: If set to true, after the embeddings have been constructed, but before the TE is estimated, the source embeddings are permuted according to our local permutation scheme.

surrogate_num_samples_ratio::AbstractFloat = 1.0: Controls the number of samples used to to construct the alternate set of history embeddings used by our local permutation scheme. This number of samples will be surrogate_num_samples_ratio * num_target_events. Corresponds to $N_{U, \textrm{surrogate}}/N_X$ in [1].

k_perm::Integer = 5: The number of neighbouring source embeddings from which to randomly select a replacement embedding in the local permutation scheme.

[1] Estimating Transfer Entropy in Continuous Time Between Neural Spike Trains or Other Event-Based Data David Shorten, Richard Spinney, Joseph Lizier bioRxiv 2020.06.16.154377; doi: https://doi.org/10.1101/2020.06.16.154377