Public Interface

CoTETE.estimate_TE_from_event_times — Function

estimate_TE_from_event_times(
  parameters::CoTETEParameters,
  target_events::Array{<:AbstractFloat},
  source_events::Array{<:AbstractFloat};
  conditioning_events::Array{<:AbstractFloat} = Float32[],
)

Estimate the TE from lists of raw event times.

Single conditioning process

Note that although the framework developed in our paper considers an arbitrary number of extra conditioning processes, at present the framework can only handle a single such process. This will change in future releases.

Examples

This example demonstrates estimating the TE between uncoupled homogeneous Poisson processes. This is covered in section II A of our paper. We first create the source and target processes, each with 10 000 events and with rate 1, before running the estimator.

julia> source = sort(1e4*rand(Int(1e4)));

julia> target = sort(1e4*rand(Int(1e4)));

julia> parameters = CoTETE.CoTETEParameters(l_x = 1, l_y = 1);

julia> TE = CoTETE.estimate_TE_from_event_times(parameters, target, source)
0.0

julia> abs(TE - 0) < 0.05 # For Doctesting purposes
true

We can also try increasing the length of the target and source history embeddings

julia> parameters = CoTETE.CoTETEParameters(l_x = 3, l_y = 3);

julia> TE = CoTETE.estimate_TE_from_event_times(parameters, target, source)
0.0

julia> abs(TE - 0) < 0.1 # For Doctesting purposes
true

Let's try some other options

julia> using Distances: Cityblock

julia> parameters = CoTETE.CoTETEParameters(l_x = 1,
                                            l_y = 2,
                                            k_global = 3,
                                            auto_find_start_and_num_events = false,
                                            start_event = 100,
                                            num_target_events = 5000,
                                            num_samples_ratio = 2.3,
                                            metric = Cityblock());

julia> TE = CoTETE.estimate_TE_from_event_times(parameters, target, source)
0.0

julia> abs(TE - 0) < 0.1 # For Doctesting purposes
true

The next example applies the estimator to a more complex problem, specifically, the process described as example B in Spinney et. al.. The application of the estimator to this example is covered in section II B of our paper. We create the source process as before. Howevever, the target process is originally created as an homogeneous Poisson process with rate 10, before a thinning algorithm is applied to it, in order to provide the dependence on the source.

julia> source = sort(1e4*rand(Int(1e4)));

julia> target = sort(1e4*rand(Int(1e5)));

julia> function thin_target(source, target, target_rate)
           # Remove target events occurring before first source
    	   start_index = 1
    	   while target[start_index] < source[1]
           	 start_index += 1
    	   end
    	   target = target[start_index:end]

	   new_target = Float64[]
    	   index_of_last_source = 1
    	   for event in target
               while index_of_last_source < length(source) && source[index_of_last_source + 1] < event
               	     index_of_last_source += 1
               end
               distance_to_last_source = event - source[index_of_last_source]
               λ = 0.5 + 5exp(-50(distance_to_last_source - 0.5)^2) - 5exp(-50(-0.5)^2)
               if rand() < λ/target_rate
               	  push!(new_target, event)
               end
           end
    	   return new_target
       end

julia> target = thin_target(source, target, 10);

julia> parameters = CoTETE.CoTETEParameters(l_x = 1, l_y = 1);

julia> TE = CoTETE.estimate_TE_from_event_times(parameters, target, source)
0.5076

julia> abs(TE - 0.5076) < 0.05 # For Doctesting purposes
true

We can also try extending the length of the target embeddings in order to better resolve this dependency

julia> parameters = CoTETE.CoTETEParameters(l_x = 3, l_y = 1);

julia> TE = CoTETE.estimate_TE_from_event_times(target, source, 3, 1)
0.5076

julia> abs(TE - 0.5076) < 0.05 # For Doctesting purposes
true

CoTETE.estimate_TE_and_p_value_from_event_times — Function

function estimate_TE_and_p_value_from_event_times(
    parameters::CoTETEParameters,
    target_events::Array{<:AbstractFloat},
    source_events::Array{<:AbstractFloat};
    conditioning_events::Array{<:AbstractFloat} = Float32[],
    return_surrogate_TE_values::Bool = false,
)

calculate the TE and the $p$ value of it being statistically different from zero.

This example demonstrates estimating the TE and $p$ value between uncoupled homogeneous Poisson processes. As the true value of the TE is zero, we expect the $p$ value to be uniformly disributed between zero and one.

We first create the source and target processes, each with 1 000 events and with rate 1, before running the estimator and the surrogate generation procedure.

Examples

julia> source = sort(1e3*rand(Int(1e3)));

julia> target = sort(1e3*rand(Int(1e3)));

julia> parameters = CoTETE.CoTETEParameters(l_x = 1, l_y = 1);

julia> TE, p = CoTETE.estimate_TE_and_p_value_from_event_times(parameters, target, source)
(0.0, 0.5)

julia> p > 0.05 # For Doctesting purposes. Should fail every now and then.
true

This second example shows using this method on coupled processes for which the true value of the TE is nonzero. As there is a strong coupling between the source and target, we expect the $p$ value to be close to 0. The application of the estimator to this example is covered in section II B of our paper. See the above examples for estimate_TE_from_event_times for more details as well as the implementation of the thinning algorithm.

We create the source process as before. Howevever, the target process is originally created as an homogeneous Poisson process with rate 10, before the thinning algorithm is applied to it, in order to provide the dependence on the source.

julia> source = sort(1e3*rand(Int(1e3)));

julia> target = sort(1e3*rand(Int(1e4)));

julia> target = thin_target(source, target, 10);

julia> parameters = CoTETE.CoTETEParameters(l_x = 1, l_y = 1);

julia> TE, p = CoTETE.estimate_TE_and_p_value_from_event_times(parameters, target, source)
(0.5, 0.01)

julia> p < 0.05 # For Doctesting purposes. Should fail very rarely.
true

CoTETE.estimate_AIS_from_event_times — Function

estimate_AIS_from_event_times(
  parameters::CoTETEParameters,
  target_events::Array{<:AbstractFloat},
  source_events::Array{<:AbstractFloat};
  conditioning_events::Array{<:AbstractFloat} = Float32[],
)

Estimate the Active Information Storage (AIS) from lists of raw event times.

See this thesis for a description of AIS.

Examples

This example estimates the AIS on an homogeneous Poisson process. The true value of the AIS on such a process is zero.

julia> target = sort(1e4*rand(Int(1e4)));

julia> parameters = CoTETE.CoTETEParameters(l_x = 1);

julia> AIS = CoTETE.estimate_AIS_from_event_times(parameters, target)
0.0

julia> abs(AIS - 0) < 0.05 # For Doctesting purposes
true

CoTETE.estimate_AIS_and_p_value_from_event_times — Function

estimate_AIS_and_p_value_from_event_times(
  parameters::CoTETEParameters,
  target_events::Array{<:AbstractFloat},
  source_events::Array{<:AbstractFloat};
  conditioning_events::Array{<:AbstractFloat} = Float32[],
)

Estimate the Active Information Storage (AIS) along with the $p$ value of the AIS being different from 0 from lists of raw event times.

See this thesis for a description of AIS.

Examples

This example estimates the AIS and $p$ value on an homogeneous Poisson process. The true value of the AIS on such a process is zero. We expect the $p$ value to be uniformly distributed between zero and 1.

julia> target = sort(1e3*rand(Int(1e3)));

julia> parameters = CoTETE.CoTETEParameters(l_x = 1);

julia> AIS, p = CoTETE.estimate_AIS_and_p_value_from_event_times(parameters, target)
(0.0, 0.5)

julia> p > 0.05 # For Doctesting purposes. Should fail from time to time
true

This next example estimates the AIS for a process where we know that the AIS must be nonzero. This process has an event occurring every one time unit, with a bit of noise added to the event times.

julia> target = sort(cumsum(ones(Int(1e3))) .+ 1e-2*randn(Int(1e3)));

julia> parameters = CoTETE.CoTETEParameters(l_x = 1);

julia> AIS, p = CoTETE.estimate_AIS_and_p_value_from_event_times(parameters, target)
(1.0, 0.01)

julia> p < 0.05 # For Doctesting purposes. Should fail from time to time
true

CoTETE.CoTETEParameters — Type

struct CoTETEParameters
    l_x::Integer = 0
    l_y::Integer = 0
    l_z::Integer = 0
    auto_find_start_and_num_events::Bool = true
    num_target_events_cap::Integer = -1
    start_event::Integer = 1
    num_target_events::Integer = 0
    num_samples_ratio::AbstractFloat = 1.0
    k_global::Integer = 5
    metric::Metric = Euclidean()
    kraskov_noise_level::AbstractFloat = 1e-8
    num_surrogates::Integer = 100
    surrogate_num_samples_ratio::AbstractFloat = 1.0
    k_perm::Integer = 5
end

l_x::Integer: The number of intervals in the target process to use in the history embeddings. Corresponds to $l_X$ in [1].

l_y::Integer: The number of intervals in the source process to use in the history embeddings. Corresponds to $l_Y$ in [1].

l_z::Integer = 0: The number of intervals in the single conditioning process to use in the history embeddings. Corresponds to $l_{Z_1}$ in [1].

Single conditioning process

Note that although the framework developed in our paper considers an arbitrary number of extra conditioning processes, at present the framework can only handle a single such process. This will change in future releases.

auto_find_start_and_num_events::Bool = true: When set to true, the start event will be set to the first event for which there are sufficient preceding events in all processes such that the embeddings can be constructed. The number of target events will be set such that all time between this first event and the last target event is included.

num_target_events_cap::Integer = -1

start_event::Integer = 1: only used when auto_find_start_and_num_events = false. The index of the event in the target process from which to start the analysis.

num_target_events::Integer = 0: only used when auto_find_start_and_num_events = false. The TE will be calculated on the time series from the timestamp of the start_event-th event of the target process to the timestamp of the start_event + num_target_events-th event of the target process.

num_samples_ratio::AbstractFloat = 1.0: Controls the number of samples used to estimate the probability density of histories unconditional of the occurrence of events in the target process. This number of samples will be num_samples_ratio * num_target_events. Corresponds to $N_U/N_X$ in [1].

k_global::Integer = 5: The number of nearest neighbours to consider in initial searches.

metric::Metric = Euclidean(): The metric to use for nearest neighbour and radius searches.

num_surrogates::Integer = 100:

surrogate_num_samples_ratio::AbstractFloat = 1.0: Controls the number of samples used to to construct the alternate set of history embeddings used by our local permutation scheme. This number of samples will be surrogate_num_samples_ratio * num_target_events. Corresponds to $N_{U, \textrm{surrogate}}/N_X$ in [1].

k_perm::Integer = 5: The number of neighbouring source embeddings from which to randomly select a replacement embedding in the local permutation scheme.

[1] Shorten, D. P., Spinney, R. E., Lizier, J.T. (2020). Estimating Transfer Entropy in Continuous Time Between Neural Spike Trains or Other Event-Based Data. bioRxiv 2020.06.16.154377.

[2] Spinney, R. E., Prokopenko, M., & Lizier, J. T. (2017). Transfer entropy in continuous time, with applications to jump and neural spiking processes. Physical Review E, 95(3), 032319.