CoTETE.jl Documentation

CoTETE.calculate_TE_from_event_times — Function

function calculate_TE_from_event_times(
    target_events::Array{<:AbstractFloat},
    source_events::Array{<:AbstractFloat},
    l_x::Integer,
    l_y::Integer;
    auto_find_start_and_num_events::Bool = true,
    start_event::Integer = 1,
    num_target_events::Integer = length(target_events) - start_event,
    num_samples_ratio::AbstractFloat = 1.0,
    k::Integer = 5,
    conditioning_events::Array{<:AbstractFloat} = [0.0],
    l_z::Integer = 0,
    metric::Metric = Euclidean(),
    is_surrogate::Bool = false,
    surrogate_num_samples_ratio::AbstractFloat = 1.0,
    k_perm::Integer = 5,
    )

Estimates the TE from lists of raw event times.

Arguments

target_events::Array{<:AbstractFloat}: A list of the raw event times in the target process. Corresponds to $X$ in [1].

source_events::Array{<:AbstractFloat}: A list of the raw event times in the source. Corresponds to $Y$ in [1].

l_x::Integer: The number of intervals in the target process to use in the history embeddings. Corresponds to $l_X$ in [1].

l_y::Integer: The number of intervals in the source process to use in the history embeddings. Corresponds to $l_Y$ in [1].

auto_find_start_and_num_events::Bool = true: When set to true, the start event will be set to the first event for which there are sufficient preceding events in all processes such that the embeddings can be constructed. The number of target events will be set such that all time between this first event and the last event is included.

start_event::Integer = 1:

num_target_events::Integer = length(target_events) - start_event:

num_samples_ratio::AbstractFloat = 1.0:

k::Integer = 5:

conditioning_events::Array{<:AbstractFloat} = [0.0]:

l_z::Integer = 0:

metric::Metric = Euclidean():

is_surrogate::Bool = false:

surrogate_num_samples_ratio::AbstractFloat = 1.0:

k_perm::Integer = 5:

warning on using auto_find_start_and_num_events

[1] Estimating Transfer Entropy in Continuous Time Between Neural Spike Trains or Other Event-Based Data David Shorten, Richard Spinney, Joseph Lizier bioRxiv 2020.06.16.154377; doi: https://doi.org/10.1101/2020.06.16.154377